create user's views delete user func

+++ app/Http/Controllers/UserController.php
@@ -7,6 +7,8 @@ use App\Imports\UsersImport;
 use App\Models\User;
 use Exception;
 use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
 use Illuminate\Database\QueryException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
@@ -16,17 +18,13 @@ use Symfony\Component\HttpFoundation\BinaryFileResponse;
 
 class UserController extends Controller
 {
-    public function index()
+    public function index(): \Illuminate\Foundation\Application|Factory|View|Application
     {
         $users = User::all();
 
         return view('pages.user-management', ['users' => $users]);
     }
 
-    public function create()
-    {
-        return view('pages.user-management-create');
-    }
 
     public function show(User $user): JsonResponse
     {
@@ -36,15 +34,30 @@ class UserController extends Controller
     public function delete($id): Application|\Illuminate\Foundation\Application|RedirectResponse|Redirector
     {
         $user = User::find($id);
-        $user->delete();
-        return redirect('/user-management')->with('success', 'Xóa người dùng thành công');
+        if ($user) {
+            $user->delete();
+            return redirect('/user-management')->with('success', 'Xóa người dùng thành công');
+        } else {
+            return redirect('/user-management')->with('error', 'Có lỗi xảy ra trong quá trình xóa người dùng');
+        }
     }
 
+    /**
+     * @throws Exception
+     */
     public function export(): BinaryFileResponse
     {
-        return Excel::download(new UsersExport, 'users.xlsx');
+        try {
+            return Excel::download(new UsersExport, 'users.xlsx');
+        } catch (\PhpOffice\PhpSpreadsheet\Writer\Exception|\PhpOffice\PhpSpreadsheet\Exception $e) {
+            throw new Exception($e->getMessage());
+        }
     }
 
+    public function create(): \Illuminate\Foundation\Application|Factory|View|Application
+    {
+        return view('pages.user-management-create');
+    }
     public function import(): RedirectResponse
     {
         $import = new UsersImport;
@@ -64,16 +77,17 @@ class UserController extends Controller
         if (count($errors) > 0) {
             $errorMessages = [];
             foreach ($errors as $error) {
-                $errorMessages[] = $error->errors()[0];
+                $errorMessages[] = $error->getMessage();
             }
             return redirect('/user-management')->with('error', implode('<br>', $errorMessages));
         }
         return redirect('/user-management')->with('success', 'Import dữ liệu từ Excel thành công');
     }
 
-    public function template()
+    public function template(): BinaryFileResponse
     {
         $file = public_path() . "/templates/import_template.xlsx";
         return response()->download($file, 'import_template.xlsx');
     }
+
 }
